feat: split patch cache to run in separate worker
