+++ src/index.ts
@@ -3,6 +3,8 @@ export { NoiseSampler } from './procgen/NoiseSampler'
 export { ProcLayer } from './procgen/ProcLayer'
 export { Biome, BlockType } from './procgen/Biome'
 export { PatchCache } from './procgen/PatchCache'
-export { LocalCache } from './procgen/LocalCache'
+export { PatchBaseCache } from './procgen/PatchBaseCache'
+export { PatchBlocksCache } from './procgen/PatchBlocksCache'
+
 // export type { MappingConf, MappingData, MappingRanges } from "./common/types"
 // export { DevHacks } from './tools/DevHacks'
